+++ lib/utils/utils.dart
@@ -285,6 +285,16 @@ class ListUtils {
     return elements;
   }
 
+  static List<T> stripNull<T>(Iterable<T?> elements) {
+    List<T> result = <T>[];
+    for (T? element in elements) {
+      if (element != null) {
+        result.add(element);
+      }
+    }
+    return result;
+  }
+
   static void add<T>(List<T>? list, T? entry) {
     if ((list != null) && (entry != null)) {
       list.add(entry);
@@ -293,6 +303,10 @@ class ListUtils {
 
   static T? first<T>(List<T>? list) {
     return ((list != null) && list.isNotEmpty) ? list.first : null;
+  }
+
+  static T? last<T>(List<T>? list) {
+    return ((list != null) && list.isNotEmpty) ? list.last : null;
   }
 
   static T? entry<T>(List<T>? list, int index) {
@@ -388,6 +402,16 @@ class SetUtils {
         set.add(entry);
       }
     }
+  }
+}
+
+class LinkedHashMapUtils {
+  static LinkedHashMap<K, V>? from<K, V>(Map<K, V>? map) {
+    return (map != null) ? LinkedHashMap<K, V>.from(map) : null;
+  }
+
+  static LinkedHashMap<K, V>? fromEntries<K, V>(Iterable<MapEntry<K, V>>? entries) {
+    return (entries != null) ? LinkedHashMap<K, V>.fromEntries(entries) : null;
   }
 }
 
@@ -954,39 +978,71 @@ class JsonUtils {
   static Map<String, dynamic>? mapValue(dynamic value) =>
     mapCastValue<String, dynamic>(value);
 
-  static Map<T, K>? mapCastValue<T, K>(dynamic value) {
-    try {
-      return (value is Map) ? value.cast<T, K>() : null;
-    }
-    catch(e) {
-      debugPrint(e.toString());
+  static Map<K, V>? mapCastValue<K, V>(dynamic value) {
+    if (value is Map) {
+      try {
+        return value.cast<K, V>();
+      }
+      catch(e) {
+        debugPrint(e.toString());
+      }
+
+      Map<K, V> result = <K, V>{};
+      value.forEach((key, value) {
+        if ((key is K) && (value is V)) {
+          result[key] = value;
+        }
+      });
+      return result;
     }
     return null;
   }
 
-  static List<T>? listValue<T>(dynamic value) {
-    try {
-      return (value is List) ? value.cast<T>() : null;
-    }
-    catch(e) {
-      debugPrint(e.toString());
+  static List<dynamic>? listValue(dynamic value) =>
+    listCastValue<dynamic>(value);
+
+  static List<T>? listCastValue<T>(dynamic value) {
+    if (value is List) {
+      try {
+        return value.cast<T>();
+      }
+      catch(e) {
+        debugPrint(e.toString());
+      }
+
+      List<T> result = <T>[];
+      for (dynamic entry in value) {
+        if (entry is T) {
+          result.add(entry);
+        }
+      }
+      return result;
+
     }
     return null;
   }
 
-  static Set<T>? setValue<T>(dynamic value) {
-    try {
-      return (value is Set) ? value.cast<T>() : null;
-    }
-    catch(e) {
-      debugPrint(e.toString());
+  static Set<dynamic>? setValue(dynamic value) =>
+      setCastValue<dynamic>(value);
+
+  static Set<T>? setCastValue<T>(dynamic value) {
+    if (value is Set) {
+      try {
+        return value.cast<T>();
+      }
+      catch(e) {
+        debugPrint(e.toString());
+      }
+
+      Set<T> result = <T>{};
+      for (dynamic entry in value) {
+        if (entry is T) {
+          result.add(entry);
+        }
+      }
+      return result;
     }
     return null;
-  }
-
-  static LinkedHashSet<T>? linkedHashSetValue<T>(dynamic value) {
-    Set<T>? set = setValue(value);
-    return (set != null) ? LinkedHashSet.from(set) : null;
   }
 
   static List<String>? stringListValue(dynamic value) {